requirements.txt
aiohttp==0.21.0
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
blessings==1.6
chardet==2.3.0
click==5.1
colorama==0.3.3
coverage==4.0.3
coveralls==1.1
cssselect==0.9.1
decorator==4.0.9
defusedxml==0.4.1
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.3
django-allauth==0.24.1
django-appconf==1.0.1
django-nose==1.4.3
django-sslify==0.2.7
django-sslserver==0.18
django-toolbelt==0.0.1
djangorestframework==3.3.2
docopt==0.6.2
falcon==0.3.0
gnureadline==6.3.3
gunicorn==19.4.5
humanize==0.5.1
ipdb==0.9.0
ipython==4.1.1
ipython-genutils==0.1.0
mypy-lang==0.4.1
nose==1.3.7
nosetests-json-extended==0.1.0
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
pilkit==1.1.13
Pillow==3.1.1
pinax-cli==0.5
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.1
PyJWT==1.4.0
pyquery==1.2.11
python-coveralls==2.6.0
python-dateutil==2.4.2
python-mimeparse==1.5.1
python-ptrace==0.8.1
python3-openid==3.0.9
PyYAML==3.11
requests==2.7.0
requests-cache==0.4.10
requests-oauthlib==0.6.1
rethinkdb==2.3.0.post3
sh==1.11
simplegeneric==0.8.1
six==1.10.0
static3==0.6.1
traitlets==4.1.0
twtxt==1.0.1
